networkSecurity/requirements.txt
python-dotenv
pandas
numpy
setuptools
Certifi
pymongo[srv]==3.12
pymongo
scikit-learn
-e. # creates package network_security.egg-info
